refactor: enhance list controls with loading state and like functionality
