+++ frontend/src/components/Views/List/controls.js
@@ -1,12 +1,12 @@
-import { useViews } from '@/stores/view'
-import { call } from 'frappe-ui'
+import { useCall } from 'frappe-ui'
 import { inject } from 'vue'
 import { useRoute } from 'vue-router'
 
 export function useControls() {
   const doctype = inject('doctype')
+  const currentView = inject('currentView')
+
   const route = useRoute()
-  const { currentView } = useViews(doctype)
 
   function createOrUpdateStandardView() {
     if (route.query.view) return
@@ -14,10 +14,11 @@ export function useControls() {
     currentView.value.doctype = doctype
     currentView.value.route_name = route.name
 
-    call(
-      'crm.fcrm.doctype.crm_view_settings.crm_view_settings.create_or_update_standard_view',
-      { view: currentView.value },
-    )
+    useCall({
+      url: '/api/v2/method/crm.fcrm.doctype.crm_view_settings.crm_view_settings.create_or_update_standard_view',
+      method: 'POST',
+      params: { view: currentView.value },
+    })
   }
 
   function updateFilter() {
@@ -32,7 +33,40 @@ export function useControls() {
     createOrUpdateStandardView()
   }
 
+  function applyRowItemFilter({ event, idx, column, item, firstColumn }) {
+    let restrictedFieldtypes = ['Duration', 'Datetime', 'Time']
+    if (restrictedFieldtypes.includes(column.type) || idx === 0) return
+    if (idx === 1 && firstColumn.key == '_liked_by') return
+
+    event.stopPropagation()
+    event.preventDefault()
+
+    let filters = currentView.value?.filters || {}
+
+    let value = item.name || item.label || item
+
+    if (value) {
+      filters[column.key] = value
+    } else {
+      delete filters[column.key]
+    }
+
+    if (column.key == '_assign') {
+      if (item.length > 1) {
+        let target = event.target.closest('.user-avatar')
+        if (target) {
+          let name = target.getAttribute('data-name')
+          filters['_assign'] = ['LIKE', `%${name}%`]
+        }
+      } else {
+        filters['_assign'] = ['LIKE', `%${item[0].name}%`]
+      }
+    }
+    updateFilter()
+  }
+
   return {
+    applyRowItemFilter,
     updateFilter,
     updateSort,
     updateColumns,
